# colblk: optimizations to PrefixBytesBuilder.Put and DetermineUintEncoding