#### colblk: add benchmark for byteWidth


#### colblk: improved byteWidth implementation

```
name              old time/op  new time/op  delta
ByteWidth/n=1-10  1.06ns ± 1%  0.63ns ± 0%  -40.84%  (p=0.001 n=7+7)
ByteWidth/n=2-10  1.05ns ± 0%  0.63ns ± 0%  -40.60%  (p=0.000 n=8+8)
ByteWidth/n=4-10  1.06ns ± 0%  0.63ns ± 0%  -40.71%  (p=0.000 n=8+7)
ByteWidth/n=8-10  1.05ns ± 0%  0.63ns ± 0%  -40.61%  (p=0.000 n=8+8)
```

#### colblk: add DetermineUintEncodingNoDelta

Add a more efficient variant used for offsets, where the minimum value
is always 0.

#### colblk: minor optimizations to PrefixBytesBuilder.Put

Reorganize the code so that we only check one `if` in the common path.
Benchmark shows a slight improvement (I expect it to be a higher when
the keys are very small).

```
PrefixBytes/alphaLen=2/building-10    710µs ± 0%   708µs ± 0%  -0.34%  (p=0.029 n=4+4)
```

#### colblk: move completedBundleLen out of sizings

This value is not used outside of `Put` so we don't need to keep two
of its values around.

----

Fixes #4322